Move SdkMeterProvider builder to top level
